feat: create base web product page
