+++ .mailmap
@@ -1,1 +1,4 @@
 Gustavo Mendonça, Software engineer <dev.gustavomendonca@protonmail.com>
+Sarah Nicacio, Frontend developer <nicaciosarah085@gmail.com>
+Rafael Raposo, Backend developer <rrafa7124@gmail.com>
+Grazielly Varanda, Backend developer <vieiravarandagrazielly@gmail.com>
