+++ apps/web/src/app/components/best-sellers/index.tsx
@@ -5,7 +5,6 @@ import { IProductCard, ProductCard } from "../product-card";
 import { cuid } from "@/utils/cuid";
 import { useMediaQuery } from "@/hooks/use-media-query";
 import { products } from "@/data/best-sellers";
-import { format } from "path";
 import { formatTextToSlug } from "@/utils/format-text-to-slug";
 import { QueryClient } from "@tanstack/react-query";
 import { api } from "@/lib/axios";
