+++ apps/web/src/app/produto/components/price-display.tsx
@@ -8,6 +8,7 @@ interface PriceProps {
   discont: boolean;
   discountAmount: number;
   isPayable: boolean;
+  isDesktop: boolean;
 }
 
 export function PriceDisplay({
@@ -17,41 +18,103 @@ export function PriceDisplay({
   discont,
   discountAmount,
   isPayable,
+  isDesktop,
 }: PriceProps) {
   return (
-    <div className="flex flex-col gap-1">
-      {discont ? (
+    <>
+      {!isDesktop ? (
+        // Mobile
         <div className="flex flex-col gap-1">
-          <span className="text-3xl text-zinc-700">
-            {convertToReal(price - discountAmount)}
-          </span>
-          <span className="text-sm text-zinc-500 line-through">
-            {convertToReal(price)}
-          </span>
-          {isPayable && (
-            <span className="text-sm text-zinc-500">
-              {numberOfInstallments}x de{" "}
-              {convertToReal(
-                calculateTotalWithInterest(numberOfInstallments, fees, price)
-              )}{" "}
-              sem juros
-            </span>
+          {discont ? (
+            <div className="flex flex-col gap-1">
+              <span className="text-3xl text-zinc-700">
+                {convertToReal(price - discountAmount)}
+              </span>
+              <span className="text-sm text-zinc-500 line-through">
+                {convertToReal(price)}
+              </span>
+              {isPayable && (
+                <span className="text-sm text-zinc-500">
+                  {numberOfInstallments}x de{" "}
+                  {convertToReal(
+                    calculateTotalWithInterest(
+                      numberOfInstallments,
+                      fees,
+                      price
+                    )
+                  )}{" "}
+                  sem juros
+                </span>
+              )}
+            </div>
+          ) : (
+            <div className="flex flex-col gap-1">
+              <span className="text-3xl text-zinc-700">
+                {convertToReal(price)}
+              </span>
+              {isPayable && (
+                <span className="text-sm text-zinc-500">
+                  {numberOfInstallments}x de{" "}
+                  {convertToReal(
+                    calculateTotalWithInterest(
+                      numberOfInstallments,
+                      fees,
+                      price
+                    )
+                  )}{" "}
+                  sem juros
+                </span>
+              )}
+            </div>
           )}
         </div>
       ) : (
         <div className="flex flex-col gap-1">
-          <span className="text-3xl text-zinc-700">{convertToReal(price)}</span>
-          {isPayable && (
-            <span className="text-sm text-zinc-500">
-              {numberOfInstallments}x de{" "}
-              {convertToReal(
-                calculateTotalWithInterest(numberOfInstallments, fees, price)
-              )}{" "}
-              sem juros
-            </span>
+          {discont ? (
+            // Desktop
+            <div className="flex flex-col gap-0.5">
+              <span className="text-3xl text-zinc-700">
+                {convertToReal(price - discountAmount)}
+              </span>
+              <span className="text-sm text-zinc-500 line-through">
+                {convertToReal(price)}
+              </span>
+              {isPayable && (
+                <span className="text-sm text-zinc-500">
+                  {numberOfInstallments}x de{" "}
+                  {convertToReal(
+                    calculateTotalWithInterest(
+                      numberOfInstallments,
+                      fees,
+                      price
+                    )
+                  )}{" "}
+                  sem juros
+                </span>
+              )}
+            </div>
+          ) : (
+            <div className="flex flex-col gap-1">
+              <span className="text-3xl text-zinc-700">
+                {convertToReal(price)}
+              </span>
+              {isPayable && (
+                <span className="text-sm text-zinc-500">
+                  {numberOfInstallments}x de{" "}
+                  {convertToReal(
+                    calculateTotalWithInterest(
+                      numberOfInstallments,
+                      fees,
+                      price
+                    )
+                  )}{" "}
+                  sem juros
+                </span>
+              )}
+            </div>
           )}
         </div>
       )}
-    </div>
+    </>
   );
 }
