+++ apps/web/src/app/produto/components/product-desktop.tsx
@@ -1,0 +1,5 @@
+export function ProductDesktop() {
+  return (
+    <h1>teste</h1>
+  );
+}
