+++ apps/web/src/components/category-divisor/index.tsx
@@ -4,7 +4,7 @@ interface CategoryDivisorProps {
 
 export function CategoryDivisor({ title }: CategoryDivisorProps) {
   return (
-    <div className="font-medium text-xl border-b-4 w-fit">
+    <div className="font-medium text-xl w-fit ">
       <h2>{title}</h2>
       <div className="mt-4 bg-amber-800 rounded w-full h-1" />
     </div>
